support for large SD cards - fixes #86

+++ src/ESP32-targz-lib.hpp
@@ -162,7 +162,7 @@
 
 // required filesystem helpers are declared outside the main library
 // because ESP32/ESP8266/RP2040 versions of <FS.h> use different abstraction flavours :)
-__attribute__((unused)) static size_t targzFreeBytesFn() {
+__attribute__((unused)) static uint64_t targzFreeBytesFn() {
   #if defined DEST_FS_USES_SPIFFS || defined DEST_FS_USES_SD || defined DEST_FS_USES_SD_MMC || defined DEST_FS_USES_LITTLEFS || defined DEST_FS_USES_PSRAMFS
     #if defined ESP32
       return tarGzFS.totalBytes() - tarGzFS.usedBytes();
@@ -184,7 +184,7 @@ __attribute__((unused)) static size_t targzFreeBytesFn() {
   #endif
 }
 
-__attribute__((unused)) static size_t targzTotalBytesFn() {
+__attribute__((unused)) static uint64_t targzTotalBytesFn() {
   #if defined DEST_FS_USES_SPIFFS || defined DEST_FS_USES_SD || defined DEST_FS_USES_SD_MMC || defined DEST_FS_USES_LITTLEFS || defined DEST_FS_USES_FFAT || defined DEST_FS_USES_PSRAMFS
     #if defined ESP32
       return tarGzFS.totalBytes();
